refactor: move jwt function to hooks folder
